+++ chirho/robust/internals/utils.py
@@ -1,0 +1,152 @@
+import functools
+from typing import (
+    Any,
+    Callable,
+    Concatenate,
+    Dict,
+    Mapping,
+    Optional,
+    ParamSpec,
+    Tuple,
+    TypeVar,
+)
+
+import pyro
+import torch
+
+from chirho.robust.ops import ParamDict
+
+pyro.settings.set(module_local_params=True)
+
+P = ParamSpec("P")
+Q = ParamSpec("Q")
+S = TypeVar("S")
+T = TypeVar("T")
+
+
+@functools.singledispatch
+def make_flatten_unflatten(
+    v,
+) -> Tuple[Callable[[T], torch.Tensor], Callable[[torch.Tensor], T]]:
+    raise NotImplementedError
+
+
+@make_flatten_unflatten.register(torch.Tensor)
+def _make_flatten_unflatten_tensor(v: torch.Tensor):
+    def flatten(v: torch.Tensor) -> torch.Tensor:
+        r"""
+        Flatten a tensor into a single vector.
+        """
+        return v.flatten()
+
+    def unflatten(x: torch.Tensor) -> torch.Tensor:
+        r"""
+        Unflatten a vector into a tensor.
+        """
+        return x.reshape(v.shape)
+
+    return flatten, unflatten
+
+
+@make_flatten_unflatten.register(dict)
+def _make_flatten_unflatten_dict(d: Dict[str, torch.Tensor]):
+    def flatten(d: Dict[str, torch.Tensor]) -> torch.Tensor:
+        r"""
+        Flatten a dictionary of tensors into a single vector.
+        """
+        return torch.cat([v.flatten() for k, v in d.items()])
+
+    def unflatten(x: torch.Tensor) -> Dict[str, torch.Tensor]:
+        r"""
+        Unflatten a vector into a dictionary of tensors.
+        """
+        return dict(
+            zip(
+                d.keys(),
+                [
+                    v_flat.reshape(v.shape)
+                    for v, v_flat in zip(
+                        d.values(), torch.split(x, [v.numel() for k, v in d.items()])
+                    )
+                ],
+            )
+        )
+
+    return flatten, unflatten
+
+
+def _flat_conjugate_gradient_solve(
+    f_Ax: Callable[[torch.Tensor], torch.Tensor],
+    b: torch.Tensor,
+    *,
+    cg_iters: Optional[int] = None,
+    residual_tol: float = 1e-10,
+) -> torch.Tensor:
+    r"""Use Conjugate Gradient iteration to solve Ax = b. Demmel p 312.
+
+    Args:
+        f_Ax (callable): A function to compute matrix vector product.
+        b (torch.Tensor): Right hand side of the equation to solve.
+        cg_iters (int): Number of iterations to run conjugate gradient
+            algorithm.
+        residual_tol (float): Tolerence for convergence.
+
+    Returns:
+        torch.Tensor: Solution x* for equation Ax = b.
+
+    Notes: This code is adapted from https://github.com/rlworkgroup/garage/blob/master/src/garage/torch/optimizers/conjugate_gradient_optimizer.py
+    """
+    if cg_iters is None:
+        cg_iters = b.numel()
+
+    p = b.clone()
+    r = b.clone()
+    x = torch.zeros_like(b)
+    z = f_Ax(p)
+    rdotr = torch.dot(r, r)
+    v = rdotr / torch.dot(p, z)
+    newrdotr = rdotr
+    mu = newrdotr / rdotr
+
+    zeros_xr = torch.zeros_like(x)
+
+    for _ in range(cg_iters):
+        not_converged = rdotr > residual_tol
+        z = torch.where(not_converged, f_Ax(p), z)
+        v = torch.where(not_converged, rdotr / torch.dot(p, z), v)
+        x += torch.where(not_converged, v * p, zeros_xr)
+        r -= torch.where(not_converged, v * z, zeros_xr)
+        newrdotr = torch.where(not_converged, torch.dot(r, r), newrdotr)
+        mu = torch.where(not_converged, newrdotr / rdotr, mu)
+        p = torch.where(not_converged, r + mu * p, p)
+        rdotr = torch.where(not_converged, newrdotr, rdotr)
+
+        # rdotr = newrdotr
+        # if rdotr < residual_tol:
+        #     break
+    return x
+
+
+def conjugate_gradient_solve(f_Ax: Callable[[T], T], b: T, **kwargs) -> T:
+    flatten, unflatten = make_flatten_unflatten(b)
+
+    def f_Ax_flat(v: torch.Tensor) -> torch.Tensor:
+        v_unflattened: T = unflatten(v)
+        result_unflattened = f_Ax(v_unflattened)
+        return flatten(result_unflattened)
+
+    return unflatten(_flat_conjugate_gradient_solve(f_Ax_flat, flatten(b), **kwargs))
+
+
+def make_functional_call(
+    mod: Callable[P, T]
+) -> Tuple[ParamDict, Callable[Concatenate[ParamDict, P], T]]:
+    assert isinstance(mod, torch.nn.Module)
+    param_dict: ParamDict = dict(mod.named_parameters())
+
+    @torch.func.functionalize
+    def mod_func(params: ParamDict, *args: P.args, **kwargs: P.kwargs) -> T:
+        with pyro.validation_enabled(False):
+            return torch.func.functional_call(mod, params, args, dict(**kwargs))
+
+    return param_dict, mod_func
